requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
pycparser==2.19
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
urllib3==1.24.1
